# cleanup RVV implementation