--- a/xxhash.h
+++ b/xxhash.h
@@ -4057,7 +4057,7 @@ XXH_PUBLIC_API XXH64_hash_t XXH64_hashFromCanonical(XXH_NOESCAPE const XXH64_can
 #  elif XXH_VECTOR == XXH_LSX   /* lsx */
 #     define XXH_ACC_ALIGN 64
 #  elif XXH_VECTOR == XXH_RVV   /* rvv */
-#     define XXH_ACC_ALIGN 64
+#     define XXH_ACC_ALIGN 64   /* could be 8, but 64 may be faster */
 #  endif
 #endif
 
@@ -5721,8 +5721,9 @@ XXH3_accumulate_512_lsx( void* XXH_RESTRICT acc,
         __m128i* const xacc    =       (__m128i *) acc;
         const __m128i* const xinput  = (const __m128i *) input;
         const __m128i* const xsecret = (const __m128i *) secret;
+        size_t i;
 
-        for (size_t i = 0; i < XXH_STRIPE_LEN / sizeof(__m128i); i++) {
+        for (i = 0; i < XXH_STRIPE_LEN / sizeof(__m128i); i++) {
             /* data_vec = xinput[i]; */
             __m128i const data_vec = __lsx_vld(xinput + i, 0);
             /* key_vec = xsecret[i]; */
@@ -5752,8 +5753,9 @@ XXH3_scrambleAcc_lsx(void* XXH_RESTRICT acc, const void* XXH_RESTRICT secret)
         __m128i* const xacc = (__m128i*) acc;
         const __m128i* const xsecret = (const __m128i *) secret;
         const __m128i prime32 = __lsx_vreplgr2vr_d(XXH_PRIME32_1);
+        size_t i;
 
-        for (size_t i = 0; i < XXH_STRIPE_LEN / sizeof(__m128i); i++) {
+        for (i = 0; i < XXH_STRIPE_LEN / sizeof(__m128i); i++) {
             /* xacc[i] ^= (xacc[i] >> 47) */
             __m128i const acc_vec = xacc[i];
             __m128i const shifted = __lsx_vsrli_d(acc_vec, 47);
@@ -5780,11 +5782,12 @@ XXH3_accumulate_512_lasx( void* XXH_RESTRICT acc,
 {
     XXH_ASSERT((((size_t)acc) & 31) == 0);
     {
+        size_t i;
         __m256i* const xacc    =       (__m256i *) acc;
         const __m256i* const xinput  = (const __m256i *) input;
         const __m256i* const xsecret = (const __m256i *) secret;
 
-        for (size_t i = 0; i < XXH_STRIPE_LEN / sizeof(__m256i); i++) {
+        for (i = 0; i < XXH_STRIPE_LEN / sizeof(__m256i); i++) {
             /* data_vec = xinput[i]; */
             __m256i const data_vec = __lasx_xvld(xinput + i, 0);
             /* key_vec = xsecret[i]; */
@@ -5814,8 +5817,9 @@ XXH3_scrambleAcc_lasx(void* XXH_RESTRICT acc, const void* XXH_RESTRICT secret)
         __m256i* const xacc = (__m256i*) acc;
         const __m256i* const xsecret = (const __m256i *) secret;
         const __m256i prime32 = __lasx_xvreplgr2vr_d(XXH_PRIME32_1);
+        size_t i;
 
-        for (size_t i = 0; i < XXH_STRIPE_LEN / sizeof(__m256i); i++) {
+        for (i = 0; i < XXH_STRIPE_LEN / sizeof(__m256i); i++) {
             /* xacc[i] ^= (xacc[i] >> 47) */
             __m256i const acc_vec = xacc[i];
             __m256i const shifted = __lasx_xvsrli_d(acc_vec, 47);
@@ -5833,13 +5837,15 @@ XXH3_scrambleAcc_lasx(void* XXH_RESTRICT acc, const void* XXH_RESTRICT secret)
 #endif
 
 #if (XXH_VECTOR == XXH_RVV)
+    #define XXH_CONCAT2(X, Y) X ## Y
+    #define XXH_CONCAT(X, Y) XXH_CONCAT2(X, Y)
 #if ((defined(__GNUC__) && !defined(__clang__) && __GNUC__ < 13) || \
         (defined(__clang__) && __clang_major__ < 16))
-    #define RVV_OP(op) op
+    #define XXH_RVOP(op) op
+    #define XXH_RVCAST(op) XXH_CONCAT(vreinterpret_v_, op)
 #else
-    #define concat2(X, Y) X ## Y
-    #define concat(X, Y) concat2(X, Y)
-    #define RVV_OP(op) concat(__riscv_, op)
+    #define XXH_RVOP(op) XXH_CONCAT(__riscv_, op)
+    #define XXH_RVCAST(op) XXH_CONCAT(__riscv_vreinterpret_v_, op)
 #endif
 XXH_FORCE_INLINE void
 XXH3_accumulate_512_rvv(  void* XXH_RESTRICT acc,
@@ -5850,37 +5856,26 @@ XXH3_accumulate_512_rvv(  void* XXH_RESTRICT acc,
     {
         // Try to set vector lenght to 512 bits.
         // If this length is unavailable, then maximum available will be used
-        size_t vl = RVV_OP(vsetvl_e64m2)(8);
+        size_t vl = XXH_RVOP(vsetvl_e64m2)(8);
 
         uint64_t*       xacc    = (uint64_t*) acc;
         const uint64_t* xinput  = (const uint64_t*) input;
         const uint64_t* xsecret = (const uint64_t*) secret;
         static const uint64_t swap_mask[16] = {1, 0, 3, 2, 5, 4, 7, 6, 9, 8, 11, 10, 13, 12, 15, 14};
-        vuint64m2_t xswap_mask = RVV_OP(vle64_v_u64m2)(swap_mask, vl);
+        vuint64m2_t xswap_mask = XXH_RVOP(vle64_v_u64m2)(swap_mask, vl);
 
-        // vuint64m2_t is sizeless.
-        // But we can assume that vl can be only 4(vlen=128) or 8(vlen=256,512)
-        size_t i=0;
-        for(; i < XXH_STRIPE_LEN / 8; i += vl){
-            /* acc_vec = xacc[i]; */
-            vuint64m2_t acc_vec = RVV_OP(vle64_v_u64m2)(xacc + i, vl);
-            /* data_vec    = input[i]; */
-            vuint64m2_t data_vec = RVV_OP(vreinterpret_v_u8m2_u64m2)(RVV_OP(vle8_v_u8m2)((const uint8_t*)(xinput + i), vl * 8));
-            /* key_vec     = secret[i]; */
-            vuint64m2_t key_vec = RVV_OP(vreinterpret_v_u8m2_u64m2)(RVV_OP(vle8_v_u8m2)((const uint8_t*)(xsecret + i), vl * 8));
-            /* data_key    = data_vec ^ key_vec; */
-            vuint64m2_t data_key = RVV_OP(vxor_vv_u64m2)(data_vec, key_vec, vl);
-            /* data_key_lo = data_key >> 32; */
-            vuint64m2_t data_key_lo = RVV_OP(vsrl_vx_u64m2)(data_key, 32, vl);
-            /* product     = (data_key & 0xffffffff) * (data_key_lo & 0xffffffff); */
-            vuint64m2_t product = RVV_OP(vmul_vv_u64m2)(RVV_OP(vand_vx_u64m2)(data_key, 0xffffffffULL, vl), data_key_lo, vl);
-            /* swap high and low halves */
-            vuint64m2_t data_swap = RVV_OP(vrgather_vv_u64m2)(data_vec, xswap_mask, vl);
-
-            acc_vec = RVV_OP(vadd_vv_u64m2)(acc_vec, product, vl);
-            acc_vec = RVV_OP(vadd_vv_u64m2)(acc_vec, data_swap, vl);
-
-            RVV_OP(vse64_v_u64m2)(xacc + i, acc_vec, vl);
+        size_t i;
+        for (i = 0; i < XXH_STRIPE_LEN/8; i += vl) {
+            vuint64m2_t data_vec = XXH_RVCAST(u8m2_u64m2)(XXH_RVOP(vle8_v_u8m2)((const uint8_t*)(xinput + i), vl * 8));
+            vuint64m2_t key_vec = XXH_RVCAST(u8m2_u64m2)(XXH_RVOP(vle8_v_u8m2)((const uint8_t*)(xsecret + i), vl * 8));
+            vuint64m2_t acc_vec = XXH_RVOP(vle64_v_u64m2)(xacc + i, vl);
+            vuint64m2_t data_key = XXH_RVOP(vxor_vv_u64m2)(data_vec, key_vec, vl);
+            vuint64m2_t data_key_hi = XXH_RVOP(vsrl_vx_u64m2)(data_key, 32, vl);
+            vuint64m2_t data_key_lo = XXH_RVOP(vand_vx_u64m2)(data_key, 0xffffffff, vl);
+            vuint64m2_t data_swap = XXH_RVOP(vrgather_vv_u64m2)(data_vec, xswap_mask, vl); /* swap high and low halves */
+            acc_vec = XXH_RVOP(vmacc_vv_u64m2)(acc_vec, data_key_lo, data_key_hi, vl);
+            acc_vec = XXH_RVOP(vadd_vv_u64m2)(acc_vec, data_swap, vl);
+            XXH_RVOP(vse64_v_u64m2)(xacc + i, acc_vec, vl);
         }
     }
 }
@@ -5890,29 +5885,23 @@ XXH_FORCE_INLINE XXH3_ACCUMULATE_TEMPLATE(rvv)
 XXH_FORCE_INLINE void
 XXH3_scrambleAcc_rvv(void* XXH_RESTRICT acc, const void* XXH_RESTRICT secret)
 {
-    XXH_ASSERT((((size_t)acc) & 63) == 0);
+    XXH_ASSERT((((size_t)acc) & 15) == 0);
     {
-        // Try to set vector lenght to 512 bits.
-        // If this length is unavailable, then maximum available will be used
-        size_t vl = RVV_OP(vsetvl_e64m2)(8);
-        uint64_t*     xacc = (uint64_t*) acc;
-        const uint64_t* xsecret = (const uint64_t*) secret;
-
-        // vuint64m2_t is sizeless.
-        // But we can assume that vl can be only 4(vlen=128) or 8(vlen=256,512)
-        size_t i = 0;
-        for (; i < XXH_STRIPE_LEN / 8; i += vl) {
-            /* xacc[i] ^= (xacc[i] >> 47) */
-            vuint64m2_t acc_vec = RVV_OP(vle64_v_u64m2)(xacc + i, vl);
-            vuint64m2_t shifted = RVV_OP(vsrl_vx_u64m2)(acc_vec, 47, vl);
-            vuint64m2_t key_vec = RVV_OP(vreinterpret_v_u8m2_u64m2)(RVV_OP(vle8_v_u8m2)((const uint8_t*)(xsecret + i), vl * 8));
-            acc_vec = RVV_OP(vxor_vv_u64m2)(acc_vec, shifted, vl);
-            /* xacc[i] ^= xsecret[i]; */
-            acc_vec = RVV_OP(vxor_vv_u64m2)(acc_vec, key_vec, vl);
-
-            /* xacc[i] *= XXH_PRIME32_1; */
-            acc_vec = RVV_OP(vmul_vx_u64m2)(acc_vec, (xxh_u64)XXH_PRIME32_1, vl);
-            RVV_OP(vse64_v_u64m2)(xacc + i, acc_vec, vl);
+        size_t count = XXH_STRIPE_LEN/8;
+        uint64_t* xacc = (uint64_t*)acc;
+        const uint8_t* xsecret = (const uint8_t *)secret;
+        size_t vl;
+        for (; count > 0; count -= vl, xacc += vl, xsecret += vl*8) {
+            vl = XXH_RVOP(vsetvl_e64m2)(count);
+            {
+                vuint64m2_t vkey = XXH_RVCAST(u8m2_u64m2)(XXH_RVOP(vle8_v_u8m2)(xsecret, vl*8));
+                vuint64m2_t vacc = XXH_RVOP(vle64_v_u64m2)(xacc, vl);
+                vuint64m2_t vsrl = XXH_RVOP(vsrl_vx_u64m2)(vacc, 47, vl);
+                vacc = XXH_RVOP(vxor_vv_u64m2)(vacc, vsrl, vl);
+                vacc = XXH_RVOP(vxor_vv_u64m2)(vacc, vkey, vl);
+                vacc = XXH_RVOP(vmul_vx_u64m2)(vacc, XXH_PRIME32_1, vl);
+                XXH_RVOP(vse64_v_u64m2)(xacc, vacc, vl);
+            }
         }
     }
 }
@@ -5920,29 +5909,38 @@ XXH3_scrambleAcc_rvv(void* XXH_RESTRICT acc, const void* XXH_RESTRICT secret)
 XXH_FORCE_INLINE void
 XXH3_initCustomSecret_rvv(void* XXH_RESTRICT customSecret, xxh_u64 seed64)
 {
-    XXH_STATIC_ASSERT((XXH_SECRET_DEFAULT_SIZE & 15) == 0);
-    XXH_STATIC_ASSERT(XXH_SEC_ALIGN == 64);
-    XXH_ASSERT(((size_t)customSecret & 63) == 0);
+    XXH_STATIC_ASSERT(XXH_SEC_ALIGN >= 8);
+    XXH_ASSERT(((size_t)customSecret & 7) == 0);
+    (void)(&XXH_writeLE64);
     {
-        const uint8_t* kSecretPtr = XXH3_kSecret;
-        size_t XXH3_kSecret_64b_len = XXH_SECRET_DEFAULT_SIZE / 16;
-        size_t i = 0;
-        for(;i < XXH3_kSecret_64b_len;) {
-            size_t vl = RVV_OP(vsetvl_e64m2)(XXH3_kSecret_64b_len - i);
+        size_t count = XXH_SECRET_DEFAULT_SIZE/8;
+        size_t vl;
+        size_t VLMAX = XXH_RVOP(vsetvlmax_e64m2)();
+        int64_t* cSecret = (int64_t*)customSecret;
+        const int64_t* kSecret = (const int64_t*)(const void*)XXH3_kSecret;
+
+#if __riscv_v_intrinsic >= 1000000
+        // ratified v1.0 intrinics version
+        vbool32_t mneg = XXH_RVCAST(u8m1_b32)(
+                         XXH_RVOP(vmv_v_x_u8m1)(0xaa, XXH_RVOP(vsetvlmax_e8m1)()));
+#else
+        // support pre-ratification intrinics, which lack mask to vector casts
+        size_t vlmax = XXH_RVOP(vsetvlmax_e8m1)();
+        vbool32_t mneg = XXH_RVOP(vmseq_vx_u8mf4_b32)(
+                         XXH_RVOP(vand_vx_u8mf4)(
+                         XXH_RVOP(vid_v_u8mf4)(vlmax), 1, vlmax), 1, vlmax);
+#endif
+        vint64m2_t seed = XXH_RVOP(vmv_v_x_i64m2)((int64_t)seed64, VLMAX);
+        seed = XXH_RVOP(vneg_v_i64m2_mu)(mneg, seed, seed, VLMAX);
+
+        for (; count > 0; count -= vl, cSecret += vl, kSecret += vl) {
+            /* make sure vl=VLMAX until last iteration */
+            vl = XXH_RVOP(vsetvl_e64m2)(count < VLMAX ? count : VLMAX);
             {
-                vuint64m2_t vlo = RVV_OP(vlse64_v_u64m2)((const uint64_t*)(const void*)kSecretPtr, 16, vl);
-                vuint64m2_t vhi = RVV_OP(vlse64_v_u64m2)((const uint64_t*)(const void*)(kSecretPtr + 8), 16, vl);
-
-                // vlo += seed64, vhi -= seed64
-                vuint64m2_t lo = RVV_OP(vadd_vx_u64m2)(vlo, seed64, vl);
-                vuint64m2_t hi = RVV_OP(vsub_vx_u64m2)(vhi, seed64, vl);
-
-                RVV_OP(vsse64_v_u64m2)((uint64_t*)customSecret, 16, lo, vl);
-                RVV_OP(vsse64_v_u64m2)((uint64_t*)(void*)((uint8_t*)customSecret + 8), 16, hi, vl);
+                vint64m2_t src = XXH_RVOP(vle64_v_i64m2)(kSecret, vl);
+                vint64m2_t res = XXH_RVOP(vadd_vv_i64m2)(src, seed, vl);
+                XXH_RVOP(vse64_v_i64m2)(cSecret, res, vl);
             }
-            i += vl;
-            kSecretPtr += vl * 16;
-            customSecret = (uint8_t*)customSecret + vl * 16;
         }
     }
 }