@camel-cdr 2025-09-07
One of the builds failed, because gcc-13 doesn't support the ratified version of the C intrinsics, only the pre-standardization version v0.11.

I've adjusted the PR to detect the intrinsics version and handle this part differently for v0.11, because gcc-13 isn't all that old and probably still worth optimizing for.