@t-mat 2025-09-08
Hi @camel-cdr , thanks for your effort.

Sorry to bother you, but could you change the following minor issues?  We should fix this in a review for the original PR.

(1) We're always using `XXH_` prefix for macro.  Therefore we should change `RVV_*` macro prefix to `XXH_RVV_*`.
(2) We keep compatibility with C90 (+unsigned long).  So we need to avoid variable declaration in `for()`.  I'm sure that no one needs C90 compiler for RVV platforms though 😄 

TODO : @t-mat, @Cyan4973 - Add `make c90test` or `make test` for QEMU platforms to detect C90 incompatibility.