@Cyan4973 2025-09-09
I'm sorry, it seems we have multiple contributions updating the same part of the code base in a short timeframe.

From a cursory glance, it seems your proposed modifications are wider than the conflicting ones.
So it's most likely worth merging even if there are a few lines of overlap.

Also, note that some recent merge are said to change performance for RISC-V, hopefully for the better.
So, it could be worth updating the benchmark numbers published in this PR.