@Cyan4973 2025-09-09
The proposed update looks generally better, respecting more strictly the conventions of this code base.

The only place where the new code does not look simpler than the one it replaces is within `XXH3_initCustomSecret_rvv()`,
but it also seems to do something more than the code it replaces.
And if that's the case, it deserves to be documented.