@camel-cdr 2025-09-09
Yes, it would remove support for clang-14/15, but those target a very early version of the intrinsics spec, I suppose it's not problematic to keep, so we should keep it for now.

Here are updated benchmarks of current dev branch vs this PR.

this PR:
```
 ===  benchmarking 1 hash functions  === 
benchmarking large inputs : from 512 bytes (log9) to 128 MB (log27) 
xxh3   ,  1375,  1585,  1702,  1784,  1811,  1830,  1753,  1795,  1833,  1842,  1835,  1826,  1829,  1825,  1823,  1817,  1812,  1806,  1805
Throughput small inputs of fixed size (from 1 to 30 bytes): 
xxh3   ,  33006753,  32378756,  32388759,  26456458,  26886618,  26461691,  26486948,  26458480,  20295073,  19775490,  19791207,  19790687,  20296864,  19792660,  19777569,  19787894,  12600821,  12228925,  12227917,  12225385,  12610281,  12219708,  12229700,  12229268,  12614733,  12229932,  12224896,  12225528,  12626099,  12223831
benchmarking random size inputs [1-N] : 
xxh3   ,  32971941,  32677008,  32600709,  29064524,  28106805,  27622978,  27318891,  27199310,  25640875,  24802593,  24241105,  23472076,  22843998,  22555019,  22244436,  22028456,  21085206,  19918033,  19650424,  18636281,  18308882,  17716234,  17294728,  17098894,  16714026,  16242514,  16193975,  15796590,  15993652,  15639239
Latency for small inputs of fixed size : 
xxh3   ,  26546537,  26331890,  26419753,  22491797,  22113224,  22389332,  22236991,  21847444,  17977422,  17443911,  17546750,  17311944,  17567202,  17499588,  17255179,  17679292,  11282087,  11268202,  11379675,  11566679,  11335494,  11331295,  11342922,  11517376,  11280604,  11270190,  11366508,  11212886,  11306947,  11458756
Latency for small inputs of random size [1-N] : 
xxh3   ,  26139882,  25449076,  25749837,  23634756,  22801857,  22483387,  22383174,  22190210,  21232498,  20483668,  20312555,  19816750,  19258911,  19053328,  18927148,  18697133,  18075905,  17219549,  17010179,  16310228,  15996902,  15637313,  15314557,  15075288,  14816993,  14434407,  14394959,  14069773,  14246875,  14021082
```

dev branch:
```
 ===  benchmarking 1 hash functions  === 
benchmarking large inputs : from 512 bytes (log9) to 128 MB (log27) 
xxh3   ,  1373,  1573,  1685,  1756,  1785,  1802,  1724,  1764,  1803,  1811,  1803,  1805,  1804,  1801,  1799,  1798,  1796,  1794,  1795
Throughput small inputs of fixed size (from 1 to 30 bytes): 
xxh3   ,  33042691,  32369134,  32363658,  26461126,  26888958,  26448156,  26466730,  26474135,  20292838,  19793708,  19796497,  19794812,  20296060,  19782713,  19788647,  19793363,  12609362,  12220713,  12224378,  12222483,  12622236,  12232776,  12229683,  12231086,  12616905,  12229987,  12238390,  12207856,  12607315,  12222307
benchmarking random size inputs [1-N] : 
xxh3   ,  33041280,  32700112,  32626082,  29418147,  28138847,  27555800,  27293902,  27005169,  25685952,  24542006,  24275296,  23497788,  22824394,  22512371,  22239976,  22029958,  21097171,  19980137,  19644445,  18635435,  18337265,  17735825,  17306719,  17084919,  16708801,  16230197,  16202913,  15790402,  15992192,  15651775
Latency for small inputs of fixed size : 
xxh3   ,  26979317,  26710596,  26837160,  22781965,  22378182,  22408914,  22507509,  22083113,  18155094,  17592255,  17704941,  17450714,  17466736,  17691670,  17573369,  17826256,  11337855,  11367778,  11443626,  11327679,  11401992,  11386482,  11388972,  11583942,  11607327,  11408812,  11438956,  11531175,  11367648,  11467378
Latency for small inputs of random size [1-N] : 
xxh3   ,  26724970,  26030854,  26068429,  23949000,  23076516,  22748765,  22611602,  22526071,  21384630,  20711765,  20479864,  19950829,  19427511,  19197238,  19056918,  18881362,  18201707,  17410772,  17093565,  16475481,  16135308,  15641262,  15362619,  15117305,  14860896,  14469501,  14427720,  14108548,  14285882,  14054207
```